backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
python-dotenv
PyJWT
passlib[bcrypt]
